api/requirements.txt
bleach
btc2fiat
boto3
ecdsa
flask
flask-cors
flask-migrate
flask-sqlalchemy
gunicorn
lnd-grpc-client>=0.3.41
lnurl
psycopg2
pycoin
pyjwt
pyqrcode
python-dateutil
python-magic
python-slugify
requests
requests_oauthlib